# download/starter-pack/openapi.yaml
openapi: 3.1.0
info:
  title: DataSphere RH Guinée API
  version: 1.0.0
  description: |
    SIRH premium SaaS multi-tenant pour la Guinée.
    Authentification JWT Bearer + header X-Tenant-ID.
    Toutes les réponses au format `{data, meta, error}`.
  contact:
    name: DataSphere RH Guinée
    email: dev@datasphererh.gn
  license:
    name: MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Local dev
  - url: https://api.datasphererh.gn/api/v1
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Connexion
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
                mfaCode: { type: string, description: "Code TOTP 6 chiffres (si 2FA actif)" }
      responses:
        '200':
          description: Succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      accessToken: { type: string }
                      refreshToken: { type: string }
                      user: { $ref: '#/components/schemas/User' }
                      requiresMfa: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '423': { description: Compte verrouillé }
...
  /employees:
    get:
      tags: [Employees]
      summary: Liste paginée des employés
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: filter[status]
          in: query
          schema: { type: string, enum: [actif, suspendu, sorti] }
        - name: filter[company_id]
          in: query
          schema: { type: string, format: uuid }
        - name: include
          in: query
          schema: { type: string, description: "Relations à inclure (company,contracts,payslips)" }
      responses:
        '200':
          description: Liste paginée
    post:
      tags: [Employees]
      summary: Crée un employé (wizard 5 étapes)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmployeeInput' }
...
  /audit/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Audit]
      summary: Détail log avec diff before/after

  # ━━━ Dashboard ━━━
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, default: 20, maximum: 100 }
...
  responses:
    Unauthorized:
      description: Non authentifié
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            title: { type: string }
            detail: { type: string }
            instance: { type: string }
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        phone: { type: string }
        roles: { type: array, items: { type: string } }
        permissions: { type: array, items: { type: string } }
        mfaEnabled: { type: boolean }
        lastLoginAt: { type: string, format: date-time, nullable: true }
...
    EmployeeInput:
      type: object
      required: [matricule, nom, prenoms, dateEmbauche, companyId]
      properties:
        matricule: { type: string }
        nom: { type: string }
        prenoms: { type: string }
        dateNaissance: { type: string, format: date }
        cnssNumero: { type: string }
        sexe: { type: string, enum: [M, F] }
        situationFamiliale: { type: string }
        nombreEnfants: { type: integer, default: 0 }
        adresse: { type: string }
        telephone: { type: string }
        email: { type: string, format: email }
        dateEmbauche: { type: string, format: date }
        companyId: { type: string, format: uuid }
...